Force building libc.
I can't think of any reason we wouldn't need it for some part of depthcharge.
In general, if we don't need it at all it will be garbage collected. Some parts
of it are needed by the trampoline code and so need prefixing. That means it,
or at least parts of it, still need to be called out seperately, but all parts
should still be built.

Change-Id: I25d9455623ad12d7b416ff414c99c3fcd8dd16c1