Bring arm64 exception handling into line with arm and x86.
Initialize the exception stack in the exception_init function. Move the
declaration of the exception stack into the exception C file instead of having
a region for it in the linker script used by all artichtectures. It was just
empty/wasted space on the other architectures.

Since I don't have a 64 bit ARM device to test this on, I just made sure it
still compiled.

Change-Id: Ifc5cce909415b86a93737ce49c30a26b12bd2a1e