Factor all path information out of engine.mk.
Define that information modularly as the Makefile.inc files are consumed.
That takes away the special meaning for the "arch" directories, for instance,
and makes them easier to handle in a generic way.

This has two main purposes. First, it avoids having a laundry list of paths
all in engine.mk, seperate from the context of why they're needed and what
they refer to.

Second, it takes us one baby step farther from having one magical version of
things. An example of that would be the current architecture, but it could
also include the linker script to use. That's necessary when, for instance,
you're working on a chip who's cores don't all implement the same architecture
(arm v4 vs. v7, for instance), or if the different components of the firmware
need to be compiled or linked differently depending on their purpose.

Subsequent steps would be to make it possible to have more than one version of
things simultaneoulsy, and to select which version to use depending on the
circumstances.

Change-Id: I365845cb2863e483680a6c8e68a7265edd4cb57f