Switch over to the "bootplan" mechanism.
A "bootplan" describes a collection of modules, and a collection of images to
build using those modules. This change adds a bootplan for running depthcharge
as a coreboot payload and configures all the remaining boards to use it. Since
image construction is handled elsewhere in that case, no images have been set
up.

Change-Id: I06d1258ae4d780c721cdceb5ec4eac6fd4916dae