kconfig: Ensure that the default config file is actually a file.
kconfig will complain if it can't open the default config file with fopen,
but it's entirely possible to open a directory with fopen, for instance the
one which actually contains the default config file. It would then proceed
without default values for variables that don't have intrinsic defaults,
causing the build to fail.

Change-Id: I1f666716416424f3080f3368d984a35b515da73e