Split the main Makefile into three parts.
One part is for interfacing with kconfig and using it to build a config.h file.
Another part is for traversing the source tree, collecting sources and building
dependencies to actually build depthcharge. The third is a wrapper which
invokes the other two.

This new structure has a couple of advantages. First, it removes the
possibility of running kconfig to set up a config manually. That was never
really something that would work like you'd expect/hope since building for a
particular board would require that certain drivers be present, for instance.

Second, this gets rid of the awkward "make defconfig" "make" sequence needed
to build depthcharge. Now you can just run "make", and it will manage the
config for you under the covers.

The idea is to shift the user's perspetive from config options to selecting
from configs as a whole, since that's what they were really doing as it was
implemented.

Additionally, this breaks Makefile a bit into modules which do a particular
job, making it a bit easier to see what's going on. Also, the top level make
file could invoke the sub-make files more than once to, for instance, build all
boards with one command.

Change-Id: I4d21995776083b6939646e841c9aef1ed16bbef8