Import coreboot's libpayload.
Depthcharge is currently dependent on this library in order to start, for some
drivers, some library code, some compiler information, and to receive
information from coreboot. Since the intention is for depthcharge to be able
to run on its own, some of the functionality in this library will need to be
separated out and/or refactored so that it can be used when not running after
coreboot. Features that aren't used at all (ncurses, for example) can be
eliminated, and code can be refactored without having to worry about upstream
acceptance or legacy compatibility.

Additionally, the libpayload build and configuration mechanisms can be
integrated with depthcharge's, psuedo-removing an external dependence and
reducing the number of independently moving parts when building a complete
image.

Change-Id: Ie148399145780003b76cb760d01ad08528518230