Collapse down to one libpayload build, and start using it.
We had been building two different libpayloads because we needed to leave GDB
out of one of them so we could have it in some cases and not in others. With
just one library to link against we can also start using our local copy and
build all bootplans/images in one shot.

Change-Id: I5c1f800af55b8e9ae26d35f7127406091ca1ce9d