Remove settings which just use the default.
Remove settings in the various configs which are just reasserting the defaults.
Subsequent changes will go through and make sure defaults are truly stable and
don't need to be pinned in individual configs. To make life slightly easier, a
couple of minor strategic defaults were added.

This change also switched libpayload over to the "defconfig" style instead of
piping 'yes ""' to make oldconfig.

Change-Id: I70db9348047aeee5e27d431b234311faf4006aea