Stop using lpgcc.
lpgcc is only appropriate when actually building as a coreboot payload. By
doing lpgcc's job directly in the depthcharge build, we can get the parts of
it that are always necessary, and optionally include the parts that are
conditionally necessary. It also removes a reason to "install" libpayload
somewhere before using it.

Change-Id: I23e5fd5052dbca802f7d3983334748158c1cb963