Build libpayload as a part of the depthcharge build.
Lump it in with the depthcharge make instead of calling a submake. It's still
being compiled as archive files.

The "libpayload" directory needed to be moved into "src" because the
depthcharge Makefile assumes that all source files start with "src" when it
figures out what the object names are going to be.

Change-Id: I92327ce3c46fdebda5e22f044848fc8af42c7a07