Clean up PCI support.
This change does a few things. First, it eliminates "libpci", a shim which
partially added support for the libpci interface to libpayload to support
running programs like flashrom as a coreboot payload. It added functions which
aren't used in depthcharge, redundant constant definitions, and some shims
which would allow using libpci's names for some libpayload PCI accessor
functions.

Second, it standardizes on one header file name and set of constants. Some
code was using one and some the other, partially based on which had the
constants it needed, and partially on which one the author likely ran across
first.

Third, it turns the #define style constants in the file into enum style
constants. That's more in line with depthcharge's philosophy of avoiding the
preprocessor and macros whenever possible.

Change-Id: I678f1308e8091ca6a2f3cab3276b23d93c220769