Clean up the endian conversion functions from libpayload.
Use the gcc builtins for 32 and 64 bit byte swapping. That will hopefully
make those functions more efficient if there are assembly instructions which
do the job.

Convert the "handy" macros for setting/clearing bits in a 32 bit register into
static inline functions. These should probably go to a different file.

Eliminate the macros defined for compatability with old code and update that
code to use the current names.

Eliminate the byte/bit streaming functions which are not used in depthcharge.

Get rid of the swab.h header file which wasn't being used.

Change-Id: I41cb99f58aa3e92463045c43fbd90dd1bf03a4ed