Move standard header files out of libpayload and into base/libc/includes.
These are distinct from the other header files in base in that they define
standard types/functions, and that they have well defined include paths which
have no relation to src/base. That's why they've been put in their own
directory which is added to the include search path. That's in contrast to
most other header files in depthcharge which are included relative to src/

Change-Id: I0b5e47676f9bc445b7c2ef1f3f4de5154dc64c78