Modify the top level Makefile so it can build multiple boards.
This will be useful for testing to make sure changes haven't broken any of
the boards, or at least prevented them from building.

Partially to support the new structure, and partially just because it should
be easier to use, the interface to the build system has changed. Rather than
specify BOARD="foo" to build board "foo", you would instead make "foo"
directly. To build multiple boards, just specify each of them as make targets.
If you want to build all known boards, you can build target "all". If you want
to know what boards are available, you can leave out the target.

The way the Makefile detects boards is it looks for directories in $(src)/board
that contain a defconfig file. That should be a pretty robust test, at least
for now. It can be revisited if the structure of board directories changes.

Change-Id: I51575153a7612f5c647960e401ba89b328cc05ca