Fold some single use constants into the code in samus's board.c.
Keeping them separate makes it harder to correlate values with where they're
used, and you still have to change the same number of places to update them.
If a value is used more than once it probably deserves a constant, but then
it should be kept local to its uses.

Change-Id: I20945e38ef7f0f83b2fbe5246ccf5ebb7573f021