Get rid of some redundancy in the path names of some i2s drivers.
For multiple files which all apply to the same device, rather than prefixing
their names, lets put them in directories. Also, for those devices and the
ones which were already in directories, take the device name out of the file
name. We already know which device it goes with based on its path.

Change-Id: Iec88bcc7b6ede1b1c9f6085ebcebd8cff48712a0