Rework how images and bootplans are hooked into the build system.
This change introduces the idea of an image which is built from various
modules using an arbitrary recipe. Rather than make each bootplan the top
level construct that depends on the image, instead use the bootplan to pick
which images to build. Then make rules define what modules to build to support
a given image, and how to put them together. This gets rid of some machinery
which attempted to build a composite bootplan object made up of modules and
images.

To avoid conflicts where two images with the same name could be built in
different ways depending on what bootplan was being used, the image recipes
are guarded with the bootplan config variable.

When being built as a coreboot module, things are slightly different because
we're not actually building the image in place, at least not yet. Instead, we
can use that mechanism to rename the modules from what depthcharge wants to
call them to what the ChromeOS build system expects them to be called.

Because make pattern rules play games with directories when applying the stem
from a target to a dependency with '%', we'll instead use $$* (and the
SECONDEXPANSION feature) to paste only the stem into the dependency without
the path munging.

Change-Id: Id3a49cc6ba06e2575705ad5599f1d285bac6f8cb