Build the trampoline as a module and eliminate the symbol prefix hack.
The previous mechanism built the source specific to the trampoline and then
prefixed its symbols, and the symbols of some basic library support it needed,
with a tag that would keep it from interlinking with the rest of depthcharge.
The sections of those object modules were then also prefixed so that they could
easily be differentiated in the linker script. The whole thing was then linked
together, with the trampoline hermetically sealed away except for its one entry
point.

The new mechanism builds the trampoline and the libraries it needs as a module
and partially links it with section garbage collection (the default) to get
rid of unnecessary code. It then makes all the symbols in the resulting ELF
local, except for the entry point and the boundaries of the stack. That ELF can
then be linked into depthcharge, and the trampoline is still isolated from the
other parts of the code.

This new implementation has several advantages. First, it's simpler than the
old one. Second, it exercises the new module infrastructure. Third, by not
prefixing symbols, it doesn't interfere with magic symbols the compiler or
linker might be using, like $a or $t or $d on ARM. depthcharge can also call
the trampoline entry point using its natural name without having to consider
the prefix.

One disadvantage of the new scheme is that when debuging with gdb, there can
be two copies of a function with the same name, for instance memcpy, and it's
not always easy to know which one you are or should be looking at. That
shouldn't be *that* much of a problem in practice, though, since there should
be very few functions in the trampoline, just load_elf, memcpy and memset on
x86 for instance.

Change-Id: I00984e1a274e5a3957ba0a0c4f446e4969dff2c7