Add externs to some power driver structure definitions.
The structures holding callbacks for some of the power drivers were statically
declared in their header file (as opposed to being dynamically allocated with
new_*), but didn't have "extern" in front of the declaration. The compiler
treated that as a "common" symbol, and combined it with the actual definition
which appeared in the .c file.

This change adds the "extern" so that we're not relying on the good graces of
the compiler/linker for things to build.

Change-Id: I4930f7295624c1dfd3b2ee304f46557268a072b2