Increase the default stack size.
I moved the stack into the bss on samus, and I was seeing a hang until I
increased the size. I suspect we were just getting lucky with it's previous
size/placement, and if it was overflowing, it was probably not by much or
onto something which caused depthcharge to fail in a noticable way.

Change-Id: If1222fb5778b0f0a563bf008e26d8a5d8d51f1f5