Work in progress on image building utility.
To avoid having one gigantic change (or at least an even more gigantic
change) which lumps too many things together, I think I'll check in this
work in progress which can build a samus image which isn't actually
complete and wouldn't boot, but has most of the right pieces in the
right places.

This tool needs to be run out of the ChromeOS build environment because
of dependencies on some ChromeOS tools and scripts. These include, but
are not necessarily limitted to; cbfstool, chromeos_version.sh,
gbb_utility, and vbutil_firmware. cbfstool and chromeos_version.sh
shouldn't be relevant in the Turquoise environment, but gbb_utility and
vbutil_firmware probably will be. Those both come from vboot which has
already separately made the transition. build_image relies on PATH being
set appropriately to find these tools.

Changes so far:

Increase the indentation from 2 spaces to 4. The ability to actually see how
much something is indented is more important than the few characters lost on
the left hand margin.

Change from being oriented around boards to being oriented around layouts.
I think there's enough generality between boards to share the majority of
their layout with only some tweaks here and there. This may need to be
revisited if there's too much divergence to manage with some simple options.

Added some constants related to FMAP flags, and enabled the "static"
flag by default. I don't know what it does, but it's set in ChromeOS, and not
setting it is a trivial difference that's easy to fix. FMAPs will likely
not be used when not running as a coreboot payload, so exactly how one
is set up will hopefully not be important in the long term.

The target board was shifted from nyan to samus.

Some things which aren't implemented yet:

IFD support. Minimally we need to duplicate ChromeOS's approach of having a
skeleton descriptor and filling that in post build. Ideally we would construct
the descriptor ourselves which would let us have a single source of truth
about region position and size.

Named constants for GBB flag values

Orient layout specific options as subcommands in argparse instead of
using parse_known_args? The layouts aren't really subcommands, but it fits
a little better than the current set up.

Add a file path configuration mechanism. Right now, there are two hardcoded
search paths for files built into the File component. That's clearly not a
workable log term solution.

Change-Id: I99ee10eb367af441a49de028c5d46570ca79a5c7