build_image: Add a component for managing the IFD.
The IFD is the Intel Firmware Descriptor, a data structure in ROM that
describes various properties of the image, including where certain parts of it
are, some permissions bits, etc., etc. The structure is consumed by the
hardware before any of our software runs and so needs to be correct even if we
don't intend to use it.

This code borrows heavily from the ifdtool which is distributed along with
coreboot.

Change-Id: I81918fa719efc9aa0a8266cc67236b1d4cd18d92