build_image: Make the "samus" layout use the IFD component.
One part of the FMAP splats across the otherwise orderly hierarchy of regions
and needs to be handled separately. All the other parts of the image are layed
out, and then that part is layered on top of the other parts. This actually
works pretty well, since it relies on size information in the IFD descriptor
blob itself instead of duplicating that information in the layout.

Change-Id: I5167b51a43e00c6dde56bd025fcdeb2c5d8d266f