build_image: Split the default place() into place() and place_children().
place() is used to recursively position and size objects in the image
hierarchy, and has historically had two jobs, recording information about the
placement of the object itself, and the directing placement of its children.
If an object needed to place its children in a non-standard way, it had to
copy and paste the other part, and then put in its custom handling.
Alternatively, the author could forget to put the other part in entirely,
leading to annoying errors.

This change splits place() into place() and place_children(). place() is
responsible for handling the placement of the object itself, primarily just
recording placement info and calling place_children, and place_children()
handles placement of the items which are contained by the current object. A
generic implementation still exists which should be applicable most of the
time, but if necessary it can be overridden, leaving the other part intact.

Change-Id: I7bbc087d69b3f47087adc9882a5ec2a561e8a122