build_image: Revamp the Area class.
This change adds some accessor functions and some sanity checking. It also
makes it possible for multiple sibling objects to have "expand" set, and to
divy up the space between them based on their expand weight which defaults to
1. Areas can now also set "shrink" which will direct build_image to size them
minimally based on the size of their contents. It also separates (for the most
part) the parameters which are set by the consuming layout, and those values
which are computed from those settings. That should hopefully make it easier
to reconstruct what's going on if something isn't working like you'd expect.

Most importantly, this also changes the algorithm used to size and place the
contents of the image. Now, instead of making one place to place and size
objects and one to write them into a buffer, there are three passes. First,
objects figure out how much space they minimally need. Second, any left over
space at a particular level of the hierarchy is distributed among any
expanding segments at that level. Finally, the contents are written to a
buffer.

Given this new policy, an Area which has an internally mandated size, for
example a File, has two options for enforcing that size. First, it can use the
"size()" function like a consumer would to set a fixed size at __init__ time.
That's best if the size you need is known by that point. Another option would
be to use "shrink()", and to return the size you want by overriding
"compute_min_size_content". That's what you should do if the size isn't
computable until the image has been set up more. For instance, the Fmap class
doesn't know how big it's going to be until it knows how many sections it
needs to describe.

Change-Id: I4cb144bff222dadb5563526e4892370b82eae0aa