build_image: Collapse componentlib and toollib.
The components.py and tools.py proxies for componentlib and toollib were set
up to make the import of their contents less verbose. Now that those all need
imagelib on the front anyway it's really not making much of a difference, and
the extra indirection is both non-standard and obfuscates what's actually
going on.

Change-Id: I47f4e2320d1e96d3fbacce6f455593cf87ff3feb