build_image: Modify how image contents are written.
Propagate fill bytes through the object hierarchy before doing any other
processing instead of having a default which can be overridden one object at
a time.

Add a Buffer class to util which helps manage a objects memory representation.
It handles some of the details of computing where exactly an object should go
in the buffer, and does some error checking which python's slice operator
doesn't do on its own. Specifically it makes sure the thing your inserting is
the same size as the space you're inserting it, and that you aren't trying to
go beyond the end of the buffer. The pad_buf function is now unused and as been
removed.

Also, the invariants for components has changed slightly. All Area based
objects that are hanging from you in the hierarchy must now be your children,
even if they don't appear in the image directly. That's so that processing can
be propagated to them without having to treat them as such a special case,
although they still aren't exactly like other children.

Also, it's now enforced that minimum size and placement (offset and final
size) are computed exactly once per object in the hierarchy. That avoids
problems where two different parents would put an object in two different
places, the second one at least partially overwriting the first. That means
that if an object needs to be consumed twice, for instance a File nees to be
in the image itself and needs to be signed in a Vblock, it needs to be
instantiated twice.

Change-Id: I8d6a6606c15930b89b56239f6ebef3edf52fd1c0