build_image: Teach Ifd about versions, and make it unlock the descriptor.
Unlocking the descriptor brings that part of the image into byte equivalency
with the cros_bundle_firmware built image. This change also teaches Ifd about
the two different versions of the IFD structures which are out there,
apparently, rather than just assuming the newest one. The one I have for samus
is actually the first version and not the newest.

Change-Id: I93f0a0ade056e2f9c707efa0c5d73d0b94a097ff