build_image: Add a DerivedArea class.
This class specializes Area for cases when the children of the Area are used
to generate its content, but aren't actually placed in the image themselves.
Examples of this sort of Area would be a Sha256 hash of some data, or the Gbb
which is built using an external utility from various files.

Change-Id: Iaae28fce7751ffad5fbd58be702226037823ef10