build_image: samus: Add a payload and a device tree to CBFS.
This brings us very close to being binary identical with the
cros_bundle_firmware image. The only differences remaining are the timestamps
in the firmware IDs, and the fact that cros_bundle_firmware adds data to the
device tree before it adds it.

By extracting an already augmented device tree from an image built by
cros_bundle_firmware, everything but the timestamps are the same. I consider
the images to now be essentially equivalent.

Change-Id: Iea7ad5d663204d146c8cfe081143b0f0f97f8d75