build_image: Stop using the FileType type for the output file name.
The FileType type seems like what you'd want in this case and sounds like it
was designed for this purpose. It has one significant drawback however, which
is that if will open a file with the default file name even if you've passed
in a different name. It will them also open the name you gave it, and return
that for you to use, leaving the first file laying around.

Change-Id: I49dd2cefe2cf98a2f471dbe542ebf1e7ab3e586c