build_image: Redo selecting a layout, and the refactor the samus layout a bit.
Layouts are now selected and configured using argparse's subparsers, which has
several advantages. For instance, he help text is more descriptive, less
error checking is required on the options detected, and less code is needed in
the layout to process the extra items.

The samus layout was rearranged a bit to get rid of samus specific paths, tidy
up how options are consumed and the options that are available, and to move
from helper functions which construct subsections of the Image to classes
which embody those subsections.

Change-Id: I649fdf383715062a1d60b93eb4162d7ef1005489