build_image: Make Area.build optionally log the current state of the image.
The log goes in a file which may or may not be stdout. That log can be used to
figure out what the build_image utility thought it was doing to help debug
layouts, or to figure out what Area a particular set of bytes belong to in the
final image.

Change-Id: If1ea024efd5f5fdd1ce389b41581140e823789ea