When using build_image, generate log output and save it to a file.
Also, change around how build_image's arguments are set up. The default
arguments are put in a variable called BUILD_IMAGE_OPTS which can be extended
to add more default arguments to build_image. The defaults include passing in
the rule target as the output file, and the target plus ".log" as the log
file.

Change-Id: I1f29755190e741b7ea82f37373cd8b6c1709c7cc