Make build_image image layout a generic config option.
That option will be necessary for any board using build_image, regardless of
what its bootplan is.

Also, this change rearranges the way build_image options are constructed in
the Makefile a bit. There are now two variables, one with options for
build_image itself, and one with options for the layout. The name of the
layout is now handled separately, and is included directly in the BUILD_IMAGE
variable.

Change-Id: Ia2d7eaf80e8552acdddea75270b0c08324b62e2a