build_image: Add a FileHarness context manager to make running tools easier.
When running an external tool, it's often necessary to marshal data into a
collection of temporary files, run the tool on those files to generate more
temporary files, read the resulting data back out, and then clean everything
up. While that's not complicated in a technical sense, it involves lots of
boilerplate, introduces the chance of leaving junk temporary files lying
around, and makes the code longer and harder to read.

This change introduces a FileHarness context manager class which takes care of
many of those details. To construct one, you need to pass in a series of data
blobs, or None. Each argument will result in one temporary file, and if it's
not None its data will be pre-loaded into the file.

The variable(s) returned by the context manager are paths for all of the
temporary files it created. When the context is exited, all the temporary
files are erased.

Change-Id: I118d584410be426d9bb45c20b3e801574d8dbfd3