build_image: Refactor Cbfs and specialize the logging.
Create a CbfsFileBase class and fold a lot of functionality into it. Also
specialize the logging of the File, Payload, and main Cbfs classes. In the case
of the payload and file classes, log various additional properties if they've
been set, and also put the CBFS file name into the Area's "name".

Also, treat the bootblock as a child object, and use Area as the base class
for the file and payload classes. The Cbfs Area is derived, but the file and
payload Areas aren't, even though they don't get pasted directly into the top
level output file.

For the Cbfs area itself, add in extra properties, and add child properties
either the base or bootblock, if one was provided, and then log a list of the
files and/or payload which are included.

Change-Id: I16a337ea9241d43bf207333d402614c6bbd481be