Actually generate header dependency files when building depthcharge.
The Makefile had the ability to consume dependency files, continuing
gracefully when none existed for, for example, a clean build. We
weren't actually generating any of them though, so they were *always* not
available, even for incremental builds. This was fine when building with
ebuilds because those are generally (though not always) done from scratch,
but it's definitely not correct and should be fixed.

Change-Id: I721abed202f53c3e5dc31f747d66ecfd34b00c4a