build_image: Add hooks which are called on Areas at different points.
These hooks will let the Areas do various house cleaning/information
gathering tasks once the hierarchy has reached different levels of completion.
The current use case for this is to gather dcdir components in the hierarchy
so that they can be entered in their containing directory. That can only
happen reliably after the whole hierarchy has been constructed, and so can't
go in the __init__ of the directory itself.

Change-Id: I25d43eb09be5737200fd23947f6ca0d389f1cc4c