build_image: Rework the samus config to use dcdir.
This is in place of CBFS, FMAP and the Index structure. The alignment of the
dcdir structures aren't guaranteed as mentioned in another change.

The legacy binary is stored in a CBFS and is slightly too big to fit in the
allocated space because of dcdir directory overhead. The legacy area could be
a raw region with no overhead, but it seems like a better idea to make
allowances for storing more than one blob of data in that area, and to be able
to show how much of the space is occupied at that moment instead of how much
space was allocated absolutely.

Coreboot's ramstage is still included in the image just to fill it out and
will be removed in the future.

Change-Id: I536808c41ca2edd6055e2863819a9894762a1e2e