dcdir: Add a dcdir utility which will help interact with dcdir based images.
The only command implemented currently is "print" which shows what's in the
image. The dcdir_structs.h header is written so that it can be used in C, and
should hopefully be shared with depthcharge in the future.

This utility is purposefully implemented to be a bit like cbfstool for two
main reasons. First, that's a structure that's been shown to work fairly well
in that other context. Second, that will make it familiar and easier to use
for people that are used to cbfstool or other tools which are similar.

Example print command output:

Found depthcharge directory structure version 1.0.

[00000000 - 00800000]
[00000000 - 00001000] +-IFD_DESC
[00001000 - 00200000] +-ME
[00200000 - 00600000] +-RW
[00200058 - 00210058] | +-MRCCACHE
[00210058 - 00214058] | +-SCRATCH
[00214058 - 00216058] | +-VPD
[00216058 - 0040b02c] | +-A
[002160a0 - 002171cc] | | +-VBLOCK
[002171cc - 002328f2] | | +-VERIFIED
[00217204 - 00227f37] | | | +-MAIN
[00227f37 - 0023288a] | | | +-REFCODE
[0023288a - 002328f2] | | | \-EC
[002328b2 - 002328d2] | | |   +-0
[002328d2 - 002328f2] | | |   \-1
[002328f2 - 0025d5a2] | | +-EC
[0023291a - 0024d87a] | | | +-0
[0024d87a - 0025d5a2] | | | \-1
[0025d5a2 - 0025d5c5] | | \-FWID
[0040b02c - 00600000] | \-B
[0040b074 - 0040c1a0] |   +-VBLOCK
[0040c1a0 - 004278c6] |   +-VERIFIED
[0040c1d8 - 0041cf0b] |   | +-MAIN
[0041cf0b - 0042785e] |   | +-REFCODE
[0042785e - 004278c6] |   | \-EC
[00427886 - 004278a6] |   |   +-0
[004278a6 - 004278c6] |   |   \-1
[004278c6 - 00452576] |   +-EC
[004278ee - 0044284e] |   | +-0
[0044284e - 00452576] |   | \-1
[00452576 - 00452599] |   \-FWID
[00600058 - 00800000] \-RO
[006000a0 - 007fbfb4]   +-GBB
[007fbfb4 - 007fffb4]   +-VPD
[007fffb4 - 007fffd7]   +-FWID
[007fffd7 - 007ffff0]   \-FIRMWARE
[007fffef - 007ffff0]     \-REAL

Change-Id: I87e25eaee74f846a1cf1e0330b439ff29fa06a96