build_image: Change how vblocks are fed what to sign.
Up to this point, vblocks wouldn't sign the actual data that was going in the
image and that was going to be verified, they signed a duplicate of that that
was constructed seperately and belonged to the vblock itself. That would
typically work, but there were a couple problems.

First, that approach is a bit inefficient. It means that, if everything goes
smoothly, the process of constructing the signed data would happen at least
twice, once to go in the image, and once to sign. The image building process
is pretty quick so that's not a big problem, but it's nice to fix.

Second, the exact circumstances of each copy couldn't be guaranteed to be
identical. For one, the placement of the two within the image would be
different. Well behaved Areas wouldn't change based on where they were, but
there's no rule against taking your absolute position into consideration while
generating your content. This might be particularly true if/when alignment
requirements come into play. The two different copies might legitimately add in
padding to obey alignment requirements, but because of their exact position,
that might make them diverge.

Now that the vblock no longer owns the data it's going to sign, it can't
pre-emptively generate a vblock to know how big it will be during the sizing
step. Instead, it generates a vblock using dummy data to see what the file
size will be, and then later generates one again using the real data which is
actually what gets used in the image. Hopefully the size of vblocks are not
dependent on their contents. I think that's a reasonable assumption to make,
and if it turns out not to be true, I don't think it would be unreasonable to
make changes to the vboot utilities to make it true.

Also, the Area which sits between the signed data and its normal parent has a
special provision to cache the data returned by "write()" so that it will be
guaranteed to be the same whether it's being called to write data into the
image or by the signing process. It should be anyway, but this will make sure.

Change-Id: I42a6f855d401423a2a329ee7dc25048204110eb3