build_image: samus: Generalize the construction of the RW area.
Generalize/systematize the layout of the RW areas in an image. There are three
different types of data stored in the RW areas. First is signed data which is
covered by a vblock in its entirity. Then there's verified data which has its
hash signed, but which is not itself signed by the vblock. Finally, there's
completely unsigned data, which at the moment is limitted to the firmware ID.

In the future there will be more stages/parts to boot, and those parts may or
may not run based on exactly what type of boot is happening. Optional
components should go in the "verified" bucket so that they can be guarunteed
to be untampered with if we need them, but they aren't loaded or hashed by
default, saving boot time.

Change-Id: Ib2db8acb50583040ca594be68d1bee7ddc6467f7