samus: Split image based bootplans for samus.
Instead of toggling between the samus_cb and samus layouts while sticking with
the same bootplan, switch back and forth between bootplans and make the layout
selected for a bootplan automatic. This will allow the bootplans to diverge
more as far as what arguments they take. It's also awkward to use a bootplan
named after coreboot which doesn't actually contain any coreboot files.

Change-Id: I71134a6f91980510aa7ad702d47829df94087f2c