build_image: samus: Generalize how ECs are plugged into an image.
This has two benefits. First, the vboot API manages ECs based on an index
and not (at least explicitly) what the EC image is used for or in. Second,
this makes the layout more general and removes one more type of board
dependence.

To support this change, the RwArea class can now handle expanding a nested
dictionary of verified/signed files. The EC images are put in a directory
called "EC" in files which are named after their index.

Change-Id: Iaf2476488edfb5c53683284c0ac41d23e6d7ffd7