imagelib: Add wrappers for objcopy, and the linker through gcc.
To avoid having to have a second or processed set of arguments for the linker,
the wrapper uses the linker through gcc like depthcharge's Makefile does.

The linking wrapper uses the CC environment variable to find a compiler,
defaulting to "gcc". It uses the flags in the environment variable LINK_FLAGS,
defaulting to an empty string.

The objcopy wrapper finds objcopy using the OBJCOPY environment variable,
defaulting to "objcopy".

Change-Id: I9dd0fc902ada5a3363f9c590a1795e63edea2cbe