fsp: Output post code 0x23 and serial char # forever on samus.
This change makes the FSP real mode entry point set up samus enough to print
characters to the serial console and output POST codes over the LPC bus. To
demonstrate that the machine is up and these mechanisms are working, it will
continuously send POST code 0x23 and the character '#' over serial, followed
by a newline and a linefeed.

Before each iteration of sending a POST code or serial character, the code
waits for "a little while" by waiting for a particular increase in the value
of the TSC. Without that, at least when only sending POST codes, the AP keeps
the EC so busy that its watchdog goes off and it reboots itself (and the AP)
over and over again. That's not a huge problem since the AP doesn't do much
and starts doing it immediately after reboot, but it means the power button
can never be pressed long enough to force the system off, and the EC reboots
before it can accept commands.

Change-Id: I4f85bf13f017067e0f91df16c5c13b3921a6051d