samus: Split the coreboot based samus config into its own board.
If you want to build a samus image which uses coreboot components, you can
build the samus-cb board. For now at least, you'd do that if you wanted a
fully functional image to boot the system and allow development on, for
instance, the kernel. You'd use the normal samus board if you wanted to work
on the firmware itself.

Change-Id: Icd203facbd3565a5ec7948e9f4784eee15808d35