modules: Make subdirectories for coreboot and fsp modules.
To avoid too many unrelated files from accumulating in one directory and to
make it easier to find things, make subdirectories for coreboot (cb) and fsp
(fsp) for modules which are specifically for those use cases. Files may be
generalized and/or moved around in the future, but at least this will create
some sort of structure to work within.

Change-Id: I2947c5c485a15413356ef0341e5f58aa6f642109