imagelib: xip: Include the .noload section in the XIP final link script.
The .noload section should stay separate from the .load section so that it
can be marked as noload and so it doesn't occupy space in the image, but it
does still need to be included in the output. Because the XIP area will be
flattened into a binary we don't actually save any space by marking things
as noload, but it's good to try to anyway.

Change-Id: I29ada3b3840a8e7e766ca1f6e6d4d41703d81da8