qemu: x86: Add some ALIGNs to the linker script.
gnu's ld is very flakey as far as how it handles alignment between sections.
If things aren't to its liking, it will just use the wrong address for things
which is not very helpful. By being conservative as far as where sections are
placed, it seems to make it do the right thing.

Change-Id: I1a482e3a531b78ca38c444816e9db488c1139915