qemu: x86: Implement loading the firmware selection module.
This change adds an ELF loader to the x86 entry module. It's used to load
and jump into the firmware selection module which is the first module
written in C and running from memory.

The loader is written in assembly, and like all the other code in the entry
module, it doesn't do any local writes. It assumes that the segments of the
ELF being loaded are in memory, even though that's not checked anywhere. That
check should be added later, once I figure out how to determine where memory
is while running under QEMU.

This change also splits up the entry module's assembly a little bit more by
moving some small header type code (a macro and a symbol definition) into
their own file so the macro can be shared with the loader.

With this change, the QEMU image will start and jump to the entry module. That
will enter protected mode, and then find and load the firmware selection
module. That module starts, but since it's still geared towards running as a
coreboot payload, it will attempt to find the sysinfo structure in memory
which will fail. It keeps executing, but the serial console will be disabled.
With some hacks to the driver, serial output can be reenabled. Future changes
will have to disentangle sysinfo from some of the drivers so they can be used
without coreboot or hacks.

Change-Id: I434ba0048b56292150c26d1d031e4b67d2bfd64a